requirements.graphics.txt
panda3d==1.10.15
panda3d-gltf==0.13
panda3d-simplepbr==0.13
